SEMANA04/CaixaEletronico/src/main/java/org/example/ContaBancaria.java
package org.example;

public class ContaBancaria {
    // Atributos da classe ContaBancaria
    private String titular;
    private int numeroConta;
    private double saldo;

    // Construtor da classe ContaBancaria
    public ContaBancaria(String titular, int numeroConta, double saldoInicial) {
        this.titular = titular;
        this.numeroConta = numeroConta;
        this.saldo = saldoInicial;
    }

    // Métodos da classe ContaBancaria
    public String getTitular() {
        return titular;
    }

    public int getNumeroConta() {
        return numeroConta;
    }

    public double getSaldo() {
        return saldo;
    }

    public void depositar(double valor) {
        if (valor > 0) {
            saldo += valor;
            System.out.println("Depósito de R$" + valor + " realizado com sucesso.");
        } else {
            System.out.println("Valor de depósito inválido.");
        }
    }

    public void sacar(double valor) {
        if (valor > 0 && valor <= saldo) {
            saldo -= valor;
            System.out.println("Saque de R$" + valor + " realizado com sucesso.");
        } else {
            System.out.println("Saque inválido. Verifique o saldo disponível.");
        }
    }

    // Método toString
    @Override
    public String toString() {
        return "ContaBancaria{" +
                "titular='" + titular + '\'' +
                ", numeroConta=" + numeroConta +
                ", saldo=" + saldo +
                '}';
    }

    // Método equals
    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;

        ContaBancaria that = (ContaBancaria) o;

        return numeroConta == that.numeroConta;
    }

    // Método hashCode
    @Override
    public int hashCode() {
        return numeroConta;
    }
}

